Add util.js.
